// src/app/models/marketing/banner-item/banner-item.ts
export class BannerItem {
    id: string;
    bannerItemName: string;
    bannerId: string;
    backgroundRGB: string;
    targetUrl: string;
    imageUrl: string;
    startDate: string;
    endDate: string;
    isDefault: boolean;
    status: number;
    statusName: string;
    priority: number;
}